feat(folio): add a headless @stll/folio/core entry and harden tree-shaking
Add a ./core subpath export pointing at a new src/core/index.ts barrel that
re-exports only the framework-neutral public API (document model, createDocx,
OOXML I/O, layout / AI / template primitives, markdown bridge). Its import
graph is React-free by construction and lint-enforced (no-react-in-core), so a
non-React adapter (Vue, a Tauri shell, a Rust core) can consume the editor core
without pulling React or @stll/ui. The root @stll/folio entry now re-exports
./core and adds the React components on top, so existing consumers are
unchanged.

Also broaden the sideEffects glob from *.css to **/*.css so nested stylesheets
(src/styles/*.css) stay marked side-effectful and are never shaken away, while
everything else stays side-effect-free for dead-code elimination.
