+++ packages/folio/src/index.ts
@@ -1,3 +1,9 @@
+// `@stll/folio` is the React adapter. It re-exports the entire headless core
+// (also importable standalone, React-free, at `@stll/folio/core`) and adds the
+// React components on top.
+
+export * from "./core";
+
 export { DocxEditor } from "./components/DocxEditor";
 export type {
   DocxEditorCollaboration,
@@ -10,149 +16,7 @@ export {
   type FormattingBarProps,
 } from "./components/FormattingBar";
 export {
-  createEmptyDocument,
-  type CreateEmptyDocumentOptions,
-} from "./core/utils/createDocument";
-export { createDocx } from "./core/docx/rezip";
-export type { Document } from "./core/types/document";
-export type { DocxCompatibility } from "./core/docx/compatibility";
-export {
-  deriveBlockId,
-  getFolioParaIdFromBlockId,
-  isFolioBlockId,
-  isSequentialFolioBlockId,
-  type DeriveBlockIdInput,
-  type FolioBlockId,
-} from "./core/types/block-id";
-
-// AI suggestion primitives — types, conflict resolution, apply, and
-// the prosemirror decoration plugin. The bar/panel UI itself lives in
-// apps/web; folio only ships the headless pieces.
-export {
-  DEFAULT_AI_SUGGESTION_PRESETS,
-  type AICitation,
-  type AICitationSource,
-  type AIChatMode,
-  type AISuggestion,
-  type AISuggestionApplyMode,
-  type AISuggestionPreset,
-  type AISuggestionSeverity,
-  type AISuggestionStatus,
-  type AIBarStatus,
-  type AIGenerateInput,
-} from "./core/ai-suggestions/types";
-export {
-  applySuggestions,
-  type ApplyResult,
-} from "./core/ai-suggestions/apply";
-export {
-  applyFolioAIEditOperations,
-  createFolioAIEditSnapshot,
-  diffWordSegments,
-  hashFolioAIBlockText,
-  normalizeFolioAIBlockText,
-  type WordDiffSegment,
-  type FolioAIBlock,
-  type FolioAIBlockAnchor,
-  type FolioAIBlockKind,
-  type FolioAIBlockPreviewRun,
-  type FolioAIComment,
-  type FolioAIEditAppliedOperation,
-  type FolioAIEditApplyMode,
-  type FolioAIEditApplyResult,
-  type FolioAIEditOperation,
-  type FolioAIEditReviewMeta,
-  type FolioAIEditSeverity,
-  type FolioAIEditSkipReason,
-  type FolioAIEditSkippedOperation,
-  type FolioAIEditSnapshot,
-  type FolioAISignatureParty,
-} from "./core/ai-edits";
-export {
-  resolveSuggestionAnchor,
-  isSuggestionStale,
-  type ResolvedAnchor,
-} from "./core/ai-suggestions/conflict";
-export {
-  buildPositionalText,
-  type PositionalText,
-} from "./core/ai-suggestions/text-positions";
-export {
-  setAISuggestionsMeta,
-  setFocusedSuggestionMeta,
-} from "./core/prosemirror/plugins/aiSuggestionDecorations";
-export { scrollFolioPositionIntoView } from "./core/paged-layout/scrollToPmPosition";
-export {
-  getFolioCaretViewportRect,
-  getFolioSelectionViewportRect,
-} from "./core/paged-layout/selectionViewportRect";
-export {
-  createAICitationDecorationsPlugin,
-  setAICitationsMeta,
-  setActiveCitationMeta,
-  type AICitationRange,
-} from "./core/prosemirror/plugins/aiCitationDecorations";
-export {
-  anonymizationDecorationsKey,
-  getAnonymizationMatches,
-  setAnonymizationTermsMeta,
-  type AnonymizationMatch,
-  type AnonymizationTerm,
-} from "./core/prosemirror/plugins/anonymizationDecorations";
-export {
-  getTemplateDirectives,
-  scanDirectives,
-  type DirectiveKind,
-  type DirectiveRange,
-} from "./core/prosemirror/plugins/templateDirectives";
-export {
-  setTemplatePreviewValues,
-  type TemplatePreviewSpan,
-  type TemplatePreviewValue,
-  type TemplatePreviewValues,
-} from "./core/prosemirror/plugins/templatePreviewValues";
-export {
-  clearTemplateSlashMenu,
-  consumeTemplateSlashQuery,
-  getTemplateSlashMenu,
-  resetTemplateSlashQuery,
-  templateSlashMenuKey,
-  type TemplateSlashMenuKeyAction,
-  type TemplateSlashMenuState,
-} from "./core/prosemirror/plugins/templateSlashMenu";
-export {
-  acceptAutocompleteSuggestion,
-  acceptAutocompleteWord,
-  appendAutocompleteToken,
-  autocompleteSuggestionKey,
-  autocompleteSuggestionPlugin,
-  clearAutocompleteSuggestion,
-  DEFAULT_AUTOCOMPLETE_DEAD_ZONE_NODES,
-  finishAutocompleteSuggestion,
-  getAutocompleteSuggestion,
-  shouldTriggerAutocomplete,
-  startAutocompleteSuggestion,
-  type AcceptAutocompleteResult,
-  type AutocompleteSuggestionPluginOptions,
-  type AutocompleteSuggestionState,
-  type AutocompleteSuggestionStatus,
-  type AutocompleteTriggerCheck,
-  type AutocompleteTriggerOptions,
-  type AutocompleteTriggerSkipReason,
-} from "./core/prosemirror/plugins/autocompleteSuggestion";
-export {
   AutocompleteCaretOverlay,
   type AutocompleteCaretOverlayProps,
   type AutocompleteCaretRect,
 } from "./paged-editor/AutocompleteCaretOverlay";
-
-// DOCX-document ↔ Markdown bridge (also available at `@stll/folio/markdown`).
-export {
-  fromMarkdown,
-  toMarkdown,
-  toMarkdownResult,
-  type ImageMeta,
-  type ImageRef,
-  type MarkdownOptions,
-  type MarkdownResult,
-} from "./core/markdown";
